Reduce launcher refactor risk with facades
Extract launcher window-argument, Chrome path discovery, and debug-port polling helpers behind the existing launcher facade as the first reviewable wave of the god-object split.

Constraint: #687 requires facade-preserving no-behavior-change waves with public imports retained.

Rejected: extracting spawn/profile/lock handling in the same PR | would broaden review risk beyond the safest Wave 1 seam.

Confidence: medium

Scope-risk: narrow

Directive: Continue #687 with separate waves; keep src/chrome/launcher.ts exports source-compatible until downstream imports migrate deliberately.

Tested: npm run build; npm run lint -- --quiet; npm run lint:tier; npm test -- tests/chrome/launcher-window-args.test.ts tests/chrome/launcher-debug-port-deadline.test.ts tests/chrome/launcher-diagnostics.test.ts tests/chrome/launcher-launch.test.ts tests/chrome/launcher-port-race.test.ts tests/chrome/launcher-restart.test.ts tests/chrome/launcher-restore-session.test.ts tests/chrome/launcher-stealth-flags.test.ts tests/chrome/launcher-zombie.test.ts --runInBand; git diff --check.

Not-tested: live Chrome manual launch smoke outside the existing launcher regression suite.
